+++ src/store/auth.ts
@@ -1,7 +1,4 @@
-import { execGitCredentialFill } from "@abstractions/gitCredentialFill";
-import { performSignInFlow } from "@abstractions/signIn";
-import * as keytarType from "keytar";
-import { commands, window } from "vscode";
+import { authentication, commands, window } from "vscode";
 import { store } from ".";
 import * as config from "../config";
 import { EXTENSION_NAME } from "../constants";
@@ -12,22 +9,6 @@ const GitHub = require("github-base");
 export function getCurrentUser() {
   return store.login;
 }
-
-export type Keytar = {
-  getPassword: typeof keytarType["getPassword"];
-  setPassword: typeof keytarType["setPassword"];
-  deletePassword: typeof keytarType["deletePassword"];
-};
-
-function getNativeKeytar(): Keytar {
-  const vscodeRequire = eval("require");
-  return vscodeRequire("keytar");
-}
-
-const keytar = getNativeKeytar();
-
-const ACCOUNT = "gist-token";
-const SERVICE = `vscode-${EXTENSION_NAME}`;
 
 const STATE_CONTEXT_KEY = `${EXTENSION_NAME}:state`;
 const STATE_SIGNED_IN = "SignedIn";
@@ -76,38 +57,14 @@ async function testToken(token: string) {
   }
 }
 
-// TODO: Support SSO when using username and password vs. just a token
-async function attemptGitLogin(): Promise<boolean> {
-  const gitSSO = config.get("gitSSO");
-  if (!gitSSO) {
-    log.info("Git SSO disabled");
-    return false;
-  }
-
-  try {
-    const token = execGitCredentialFill();
-    if (token && token.length > 0 && (await testToken(token))) {
-      log.info("Git SSO succeeded");
-      await keytar.setPassword(SERVICE, ACCOUNT, token);
-      return true;
-    }
-
-    return false;
-  } catch (e) {
-    log.info(`Git SSO failed: ${e.nessage}`);
-    return false;
-  }
-}
-
 const TOKEN_RESPONSE = "Enter token";
 export async function ensureAuthenticated() {
-  const password = await getToken();
-  if (password) {
+  if (await getToken()) {
     return;
   }
 
   const response = await window.showErrorMessage(
-    "You need to sign-in with GitHub to perform this operation.",
+    "You need to sign-in with GitHub to perform this action.",
     TOKEN_RESPONSE
   );
   if (response === TOKEN_RESPONSE) {
@@ -115,38 +72,12 @@ export async function ensureAuthenticated() {
   }
 }
 
-async function deleteToken() {
-  await keytar.deletePassword(SERVICE, ACCOUNT);
-}
+export async function getToken(scopes: string[] = [GIST_SCOPE]) {
+  const session = await authentication.getSession("github", scopes, {
+    createIfNone: false
+  });
 
-export async function getToken() {
-  const token = await keytar.getPassword(SERVICE, ACCOUNT);
-  return token;
-}
-
-export async function initializeAuth() {
-  markUserAsSignedOut();
-
-  const isSignedIn = await isAuthenticated();
-  if (isSignedIn) {
-    log.info("Signed in, checking for token's validity...");
-    const currentToken = (await getToken())!;
-    const tokenStillValid = await testToken(currentToken);
-    if (!tokenStillValid) {
-      log.info("Clearing token, since it is no longer valud");
-      await deleteToken();
-      return;
-    }
-  } else {
-    log.info("Not signed in, attempting git SSO...");
-    const gitSSO = await attemptGitLogin();
-    if (!gitSSO) {
-      return;
-    }
-  }
-
-  log.info("Marking user as signed in");
-  await markUserAsSignedIn();
+  return session?.accessToken;
 }
 
 export async function isAuthenticated() {
@@ -162,21 +93,12 @@ async function markUserAsSignedIn() {
 }
 
 export async function signIn() {
-  const token = await performSignInFlow();
+  if (await performSignInFlow()) {
+    window.showInformationMessage(
+      "You're successfully signed in and can now manage your GitHub gists and repositories!"
+    );
 
-  if (token) {
-    if (await testToken(token)) {
-      window.showInformationMessage(
-        "You're successfully signed in and can now manage your GitHub gists and repositories!"
-      );
-
-      await keytar.setPassword(SERVICE, ACCOUNT, token);
-      await markUserAsSignedIn();
-    } else {
-      window.showErrorMessage(
-        "The specified token isn't valid or doesn't inlcude the gist scope. Please check it and try again."
-      );
-    }
+    await markUserAsSignedIn();
   }
 }
 
@@ -187,6 +109,5 @@ function markUserAsSignedOut() {
 }
 
 export async function signout() {
-  await deleteToken();
   markUserAsSignedOut();
 }